feat(human-context): Add stewardship allocations and contributor presences
Human Context Sprint - Enable human-centric content stewardship:

Database (elohim-storage):
- Add stewardship_allocations table with allocation ratios, governance states
- Add contributor_presences table for identity lifecycle tracking
- Add human_relationships, economic_events, content_mastery tables
- Implement Rust CRUD operations with bulk allocation support
- Add REST API endpoints for all new tables

TypeScript Layer:
- StewardshipAllocation model with wire format transformers
- StewardshipAllocationService for allocation management
- StorageApiService extensions for new endpoints
- MasteryService, RelationshipService, EventService

UI Components:
- CreatePresenceComponent (/identity/presences/create)
  - Create contributor presences for absent contributors
  - External identifiers (GitHub, ORCID, email)
  - Establishing content typeahead
- StewardshipDashboardComponent (/identity/stewardship)
  - View stewardship portfolio with allocation ratios
  - Recognition accumulated tracking
  - Active disputes indicator

Session Migration:
- Integrate mastery migration into account creation flow
- Auto-merge localStorage progress when users register

Bootstrap Seeding:
- Matthew Dowell presence seed data as protocol founder
- seed-stewardship.ts script for bulk allocation bootstrapping

Philosophy (from Manifesto Part IV-C):
"Content isn't ever owned by who might create it, it's stewarded by
whoever has the most relational connection to the content itself."

+++ holochain/elohim-storage/src/db/content_diesel.rs
@@ -34,7 +34,7 @@ pub struct CreateContentInput {
     #[serde(default)]
     pub blob_cid: Option<String>,
     #[serde(default)]
-    pub content_size_bytes: Option<i64>,
+    pub content_size_bytes: Option<i32>,
     #[serde(default)]
     pub metadata_json: Option<String>,
     #[serde(default = "default_reach")]
